# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kasane",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.3.2"
  },
  "paths": {
...
    "/databases": {
      "get": {
        "tags": [
          "Databases"
        ],
        "summary": "データベース一覧の取得",
        "description": "ユーザー権限に応じて、アクセス可能なデータベースの一覧を取得します。\n\n- **グローバル管理者**: システム内の全データベースが見えます。\n- **一般ユーザー**: 自分が権限を持っているデータベースだけが見えます。",
        "operationId": "database_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseInfoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Databases"
        ],
        "summary": "データベースの作成",
        "description": "新しいデータベースを作成します。この操作は `global` スコープの `manage` 以上のロールが必要です。",
        "operationId": "database_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    },
    "/databases/{db_name}/tables": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "テーブル一覧の取得",
        "description": "指定したデータベース内に存在するテーブルのうち、呼び出したユーザーが Read 以上の権限を\n持つものだけを返します。テーブル単位の権限しか持たないユーザーも、そのテーブルだけが\n見える形で一覧を取得できます。",
        "operationId": "table_list",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "テーブルの作成",
        "description": "指定したデータベース内に新しいテーブルを作成します。この操作はデータベースのManage以上の権限が必要です。",
        "operationId": "table_create",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "リクエストが不正（パラメータエラーなど）"
          },
          "409": {
            "description": "同名のテーブルが既に存在する"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/databases/{db_name}/tables/{table_name}/data/search": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "データの一括取得",
        "description": "空間IDと値をJSON配列として一括で取得します。この操作はデータベースのRead以上の権限が必要です。",
        "operationId": "data_get",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          },
          {
            "name": "format",
            "in": "query",
            "description": "出力フォーマット(singleId, rangeId, flexId)",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OutputFormat"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "最大取得件数",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "テーブルが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/databases/{name}": {
      "get": {
        "tags": [
          "Databases"
        ],
        "summary": "データベース情報の取得",
        "description": "指定したデータベースの詳細情報を取得します。対象データベースのRead以上の権限が必要です。",
        "operationId": "database_info",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Databases"
        ],
        "summary": "データベースの削除",
        "description": "指定したデータベースを削除します。この操作はGlobal Admin権限が必要です。",
        "operationId": "remove_database",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Databases"
        ],
        "summary": "データベースの更新",
        "description": "指定したデータベースの名前や説明を変更します。対象データベースのManage以上の権限が必要です。",
        "operationId": "database_update",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "400": {
            "description": "リクエストが不正（パラメータエラー、または新しいデータベース名が不正）"
          },
          "403": {
            "description": "権限不足（Manage権限がない場合）"
          },
          "404": {
            "description": "変更元のデータベースが存在しない"
          },
          "409": {
            "description": "変更後のデータベース名がすでに存在する"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/databases/{name}/copy": {
      "post": {
        "tags": [
          "Databases"
        ],
        "summary": "データベースのコピー",
        "description": "指定したデータベースをコピーします。この操作はGlobal Admin権限が必要です。",
        "operationId": "database_copy",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "コピー元データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "source_database"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "リクエストが不正（パラメータエラー、または不正なデータベース名）"
          },
          "403": {
            "description": "権限不足（Global Admin権限がない場合）"
          },
          "404": {
            "description": "コピー元データベースが存在しない"
          },
          "409": {
            "description": "コピー先データベース、またはコピー元と同名のデータベースがすでに存在する"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "global_admin"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CopyDatabaseRequest": {
        "type": "object",
        "required": [
          "copy_name"
        ],
        "properties": {
          "copy_name": {
            "type": "string",
            "example": "example_database_copy"
          }
        }
      },
...
      "CreateDatabaseRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "example_database"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "データベースの説明文"
          }
        }
      },
      "CreateTableRequest": {
        "type": "object",
        "description": "時空間IDと値が対応するTableを作成する",
        "required": [
          "name",
          "data_type",
          "max_zoom_level"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "example_table"
          },
          "data_type": {
            "$ref": "#/components/schemas/TableDataType"
          },
          "max_zoom_level": {
            "type": "integer",
            "format": "int32",
            "example": 25,
            "maximum": 30,
            "minimum": 0
          },
          "constraints": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TableConstraints",
                "description": "値に対する制約。指定しない場合は制約なしとなる。"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "テーブルの説明です"
          }
        }
      },
...
      "DatabaseInfoResponse": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "example_database"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "データベースの説明文"
          }
        }
      },
...
      "GetDataRequest": {
        "type": "object",
        "description": "空間IDの配列を指定して値を取得する",
        "required": [
          "spatial_ids"
        ],
        "properties": {
          "spatial_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpatialId"
            }
          },
          "zoom_level_policy": {
            "$ref": "#/components/schemas/ZoomLevelPolicy"
          }
        }
      },
      "GetDataResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/GetDataResponseSingle"
          },
          {
            "$ref": "#/components/schemas/GetDataResponseRange"
          },
          {
            "$ref": "#/components/schemas/GetDataResponseFlex"
          }
        ]
      },
      "GetDataResponseFlex": {
        "type": "object",
        "required": [
          "dictionary",
          "data"
        ],
        "properties": {
          "dictionary": {
            "type": "array",
            "items": {}
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataGroup_FlexId"
            }
          }
        }
      },
      "GetDataResponseRange": {
        "type": "object",
        "required": [
          "dictionary",
          "data"
        ],
        "properties": {
          "dictionary": {
            "type": "array",
            "items": {}
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataGroup_RangeId"
            }
          }
        }
      },
      "GetDataResponseSingle": {
        "type": "object",
        "required": [
          "dictionary",
          "data"
        ],
        "properties": {
          "dictionary": {
            "type": "array",
            "items": {}
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataGroup_SingleId"
            }
          }
        }
      },
...
      "OutputFormat": {
        "type": "string",
        "enum": [
          "singleId",
          "rangeId",
          "flexId"
        ]
      },
...
      "SingleId": {
        "type": "object",
        "description": "通常の時空間IDを表す型。\n\n時間を指定する場合は `i` と `t` を必ず両方指定すること。\n両方省略した場合は「全時間」を表す。",
        "required": [
          "z",
          "f",
          "x",
          "y"
        ],
        "properties": {
          "z": {
            "type": "integer",
            "format": "int32",
            "description": "ズームレベル。",
            "example": 20,
            "maximum": 30,
            "minimum": 0
          },
          "f": {
            "type": "integer",
            "format": "int32",
            "description": "高さ方向のインデックス。",
            "example": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "description": "経度方向のインデックス。",
            "example": 931386,
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "description": "緯度方向のインデックス。",
            "example": 412905,
            "minimum": 0
          },
          "i": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "時間間隔（単位: 秒）。1970-01-01T00:00:00Z を起点とする。任意の値は指定できず、 次の暦の単位のいずれかのみ指定できる: `1`、`60`、`3600`、`86400`、`34359738368`（システムの全時間）。",
            "example": 3600,
            "maximum": 34359738368,
            "minimum": 1
          },
          "t": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "時間インデックス。",
            "example": 1,
            "minimum": 0
          }
        }
      },
      "SpatialId": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SingleId"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "singleId"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RangeId"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rangeId"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FlexId"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "flexId"
                    ]
                  }
                }
              }
            ]
          }
        ]
      },
...
      "TableConstraints": {
        "oneOf": [
          {
            "type": "object",
            "description": "`Text` 型に対する制約。",
            "required": [
              "type"
            ],
            "properties": {
              "min_length": {
                "type": [
                  "integer",
                  "null"
                ],
                "description": "最小文字数",
                "minimum": 0
              },
              "max_length": {
                "type": [
                  "integer",
                  "null"
                ],
                "description": "最大文字数",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "Text"
                ]
              }
            },
            "example": {
              "max_length": 100,
              "min_length": 1,
              "type": "Text"
            }
          },
          {
            "type": "object",
            "description": "`Int` 型に対する制約。",
            "required": [
              "type"
            ],
            "properties": {
              "min": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "description": "最小値（境界値を含む）"
              },
              "max": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "description": "最大値（境界値を含む）"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Int"
                ]
              }
            },
            "example": {
              "max": 100,
              "min": 0,
              "type": "Int"
            }
          },
          {
            "type": "object",
            "description": "`Float` 型に対する制約。",
            "required": [
              "type"
            ],
            "properties": {
              "min": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "double",
                "description": "最小値（境界値を含む）"
              },
              "max": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "double",
                "description": "最大値（境界値を含む）"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Float"
                ]
              }
            },
            "example": {
              "max": 100.0,
              "min": 0.0,
              "type": "Float"
            }
          },
          {
            "type": "object",
            "description": "`Enum` 型に対する制約。",
            "required": [
              "choices",
              "type"
            ],
            "properties": {
              "choices": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "許容される文字列の配列"
              },
              "mapping": {
                "type": "object",
                "description": "内部用: 文字列から整数IDへのマッピング",
                "additionalProperties": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "next_id": {
                "type": "integer",
                "format": "int32",
                "description": "内部用: 次に割り当てる整数ID",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "Enum"
                ]
              }
            },
            "example": {
              "choices": [
                "red",
                "blue",
                "green"
              ],
              "type": "Enum"
            }
          }
        ],
        "description": "テーブルの値に対する制約。\n\n`type` フィールドで制約の種類を指定する。指定できる種類は対象のデータ型に依存する。\n- `Text`: 文字列の長さ制約（`min_length`, `max_length`）\n- `Int` / `Float`: 数値の範囲制約（`min`, `max`）\n- `Enum`: 許容される選択肢の制約（`choices`）\n\n`Presence` および `Boolean` には制約を指定できない。"
      },
      "TableDataType": {
        "type": "string",
        "description": "Table内の時空間IDに付与する値の型。",
        "enum": [
          "Text",
          "Int",
          "Float",
          "Boolean",
          "Enum",
          "Presence"
        ]
      },
...
      "TableListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/TableSummary"
        }
      },
      "TableSummary": {
        "type": "object",
        "description": "テーブルの基本情報。一覧で返す軽量ビューで、件数は含まない。",
        "required": [
          "name",
          "data_type",
          "max_zoom_level"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "example_table"
          },
          "data_type": {
            "$ref": "#/components/schemas/TableDataType"
          },
          "max_zoom_level": {
            "type": "integer",
            "format": "int32",
            "example": 25,
            "minimum": 0
          },
          "constraints": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TableConstraints"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "テーブルの説明文です"
          }
        }
      },
      "UpdateDatabaseRequest": {
        "type": "object",
        "properties": {
          "new_name": {
            "type": [
              "string",
              "null"
            ],
            "example": "example_database_renamed"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "更新されたデータベースの説明です"
          }
        }
      },
...
      "ZoomLevelPolicy": {
        "type": "string",
        "description": "テーブルの max_zoom_level よりも大きなズームレベル値の空間IDが入力された場合の処理方針",
        "enum": [
          "Error",
          "Ignore",
          "Normalize"
        ]
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}